Add files via upload
This is the base code needed to allow the bot to exist in a discord server. I also added the code to allow the user to provide the desired input for the chocolate bar problem. I'm envisioning the function for solving the chocolate bar problem will exist in a separate file (chocolate.py or something) which is then included in this one.