Reorganizing ReMBot's base code + added async framework
Changed around the organization to make it more scale-able. As it was, it would've become messy very quickly. I also added the framework for running actions concurrently. This is all yet to be debugged since I'm having some issues getting ReMBot to recognize the server programmatically. I'm continuing work on this but wanted to upload some code to have it visible.