Complete 'customer' useCase handling for ReMBot
Implemented full dialogue between user and ReMBot. User specifies that they are a customer. Then, ReMBot queries the user for the name of a customer node in the provided graph. If the user can't provide the name properly after 3 tries, break. Else, call the shortest-path solver and then tell the customer the name of the factory they should order from as well as the total cost of shipping from that factory.